+++ pywb/apps/frontendapp.py
@@ -271,10 +271,16 @@ class FrontEndApp(object):
 
         self.setup_paths(environ, coll, record)
 
-        wb_url_str = to_native_str(url)
+        request_uri = environ.get('REQUEST_URI')
+        script_name = environ.get('SCRIPT_NAME', '') + '/'
+        if request_uri and request_uri.startswith(script_name):
+            wb_url_str = request_uri[len(script_name):]
 
-        if environ.get('QUERY_STRING'):
-            wb_url_str += '?' + environ.get('QUERY_STRING')
+        else:
+            wb_url_str = to_native_str(url)
+
+            if environ.get('QUERY_STRING'):
+                wb_url_str += '?' + environ.get('QUERY_STRING')
 
         metadata = self.get_metadata(coll)
         if record:
@@ -396,6 +402,7 @@ class FrontEndApp(object):
 
     def init_proxy(self, config):
         proxy_config = config.get('proxy')
+        self.proxy_prefix = None
         if not proxy_config:
             return
 
@@ -423,11 +430,23 @@ class FrontEndApp(object):
         else:
             logging.info('Proxy enabled for collection "{0}"'.format(proxy_coll))
 
-        prefix = '/{0}/bn_/'.format(proxy_coll)
+        if proxy_config.get('use_head_insert', True):
+            self.proxy_prefix = '/{0}/bn_/'.format(proxy_coll)
+        else:
+            self.proxy_prefix = '/{0}/id_/'.format(proxy_coll)
 
-        self.handler = WSGIProxMiddleware(self.handle_request, prefix,
+        self.handler = WSGIProxMiddleware(self.handle_request,
+                                  self.proxy_route_request,
                                   proxy_host=proxy_config.get('host', 'pywb.proxy'),
                                   proxy_options=proxy_config)
+
+    def proxy_route_request(self, url, environ):
+        """ Return the full url that this proxy request will be routed to
+        The 'environ' PATH_INFO and REQUEST_URI will be modified based on the returned url
+
+        Default is to use the 'proxy_prefix' to point to the proxy collection
+        """
+        return self.proxy_prefix + url
 
 
 # ============================================================================